reorga architecture
